프로덕트 매니저 컴포넌트 변경
